Change: Configuration Ini parser to only use parse_ini_string() in StrictMode
eZ Publish has lots of ini files that do not conform to the strictnes of how php's ini parser works.
So from now on there is no parse_ini_string use first and then fallback to ezcConfigurationIniReader.
Former is only used if in strictMode, and latter is used if not (default).

Strict mode can be enabled with the following global config (config.php):

    $globalConfiguration['base']['configuration']['ini-parser']['strict'] = true;