Implemented: Fetching of full content in search.
The search handler now fetches the full content objects found. The
implementation re-uses the query builder used to load single content
objects and the corresponding mapper. Due to the changed fetch, 2 test
cases needed to be adjusted in terms of sort order. This is irrelevant
for now, since sorting is not defined anyway.