Add: isLoaded() on Lazy collection
Can be moved to some interface at some point and reuse in proxies.
Needed for servide layer to make sure collection values are not appended and at the same time triggering collection load causing duplicate entities to end up in collection. This is however caused by not having identity map in service layer, and the other reason for adding this is optimization: do not load the collection if you don't have to.