Implement: Translatable interface
Based on ideas that came up this sprint when talking to ls and oms to have such an api
on all classes that are translatable.
For instance if this is added on Content\Type, then the idea is that you setLangauge()
and then use ->name and ->description like simple strings, while there is logic underneath
that maps it to current language. If propertie does not exist in current language, then
it should fallback to language as defined in settings / always available language flag.