Add: Simple clone function on Model for internal use in unit tests
Warning: Does not take care about lots of dependant internal objects. So cloning is
not supported in any way atm and striclty meant for internal (unit test) use.
A lot of work will have to be put into specifc clone functionality for all domain objects if cloning is ever going to be a supported way of copying a object.
So the addintional complexity and amount of code needed is probably not worth it compared to having specifc copy API's.