Added support for lazy loaded services (Prefix: '%') using closures
Also added support for callable dependencies (variables, services and lazyloaded services) using :: syntax.
Also removed support for using $serviceContainer as dependencie, should hopefully not be needed now with the lazy loading and callable support.

One downside with the lazy loading feature: Closures in PHP does not inherit the scope they where created in in 5.3, so not possible to call internalGet() to be able to use private services as lazyloaded services.