Changed use of  SPI/../Query/ to API/../Query/ & removed SPI variant
This commit should specifically be reviewed in the change made in:
eZ/Publish/Core/Persistence/Legacy/Content/Search/Gateway/CriterionHandler/Field.php

API version of Query only takes fieldIdentifier to allow people to filter across several contentTypes if they want to.
The question here is not DB performance, but if the change is ok and if we should add additional unit test for this new use case.