Fix inconsistent issues in data.json and use of persistence api
InMemory makes it easy to check a central place if objects have all values filled in or not, both in output and inout.
Testing this revaled that a lot of tests does not bother to set all properties, but it is rather in service layer we
need to enforce that properties has a value other then null.