Implemented: Integration tests authenticator.
The IntegrationTestAuthenticator is a special authentication mechanism,
which is only used for the integration tests. It submits the currently
set user ID to the server, in a special header, so that a special
authentication module on the back end can receive it and set it
according to the test case. In addition, this special HTTP client now
handles the test case session management.