Added new dump files from 4.7 install with ezdemo and ezflow
These dumps are w/o demo content, as the idea is to make sure unit tests doens't rely on demo content.
Tests should be migrated to use ezdemo, as that is a newer package, but ezflow one provided as a inbetween step for checking unit tests for use of demo content before switching to ezdemo and adjusting for changes in dump.

Idea is that unit tests should