Fixed: Re-use database connection and schema in tests.
After removing the caching language handler (legacy persistence), the DB
connection can now be re-used in the test setup and the database schema
does not need to be re-applied in every test case. This does not only
save time when running the test suite, but also allows running against
MySQL for the first time.