Initial version of simple object cache
Simple version for testing, alpha and further discussion.
It is kept simple to allow it to be fast, and it stores only 10 objects
pr class to avoid high memory use.

Mainly to avoid that repetetive load calls during one or several calls
to the api for same object results in many spi calls.
Beacuse permissions needs content object, only Content and ContentInfo
api's are currently cached.

@todo: Add cache for loadLocation, but problem here is that you need to maintain
relationship between Content and Location, this would have been much simpler in
spi as Location is property of Content, imo how it should have been in API as well.