REST: Implemented POST /content/views
Partial implementation only.

Some Criterion parsers don't exist yet:
- FieldCritierion
- DateMetaDataCritierion
- VersionStatusCriterion
- URLAliasCriterion
- UserMetaDataCriterion

The visitor is incomplete as well, and will only output the view identifier + the results.
In the results, the result count that is displayed is incorrect, as well as the index.

Note that in any case, this implementation does NOT save the view. It only executes it.