Potential approach for changing ezp-next to use Composer and Symfony Container
!! DO NOT MERGE: this currently breaks all tests !!

Not sure if this is the way it should be done, but here is an initial commit
for changing boostrap.php and contriner.php to stop using custom autoload,
configuration, continer and ini settings, syncing with the approach in eZ Publish 5.

This approach makes sure booststrap and container works both in the context of API
(ezp-next) as standalone, and also witin ezpublish5 context.
Meaning there is no longer a need to install vendors in both eZ Publish 5 and API when
in ezpublish5 context (builds).

This is purly for starting an discussion on how we can expose container to command line
scripts and integration/unit tests (they need to be able to get Reposiotory object).

If this approach is taken, then the following remains:
- cleanup /  change tests to not check for legacyKernel env variable but instead get it via injection
- default settings might need to be moved from ezpublish5 to API (ezp-next)
- Service tests that manipulates the contaner needs to be modified
- (..)