Fix EZP-20281 ezcBaseValueException when executing setup commands
Short:
Move lazy Repository factory out of the main Reposiotory factory as this one depends on
several sevices being setter injected.

Long:
This solves the issue because some of the console commands in Symfony are loading router
during init to check it's instance (all router console commands), and our Url alias
router depends on LazyRepository, and the main Reposiotry Factory depends eagerly on several
of Reporitory dependencies being loaded true setter injection (compiler pass in this case).
And one of them (Permission limitations) depends on Persistence, which again depends on dsn,
which again dependas on proper siteaccess being set, which is not the case yet.

In essence; by lazy loading this, persistence is not at all, and also not loaded with a
dbParam value set to null because siteaccess is not set yet.