EZP-21990: Added FieldStorage events to SPI
Added handleEvent to FieldStorage SPI

FieldStorageEvents are triggered in ContentService

Since nodes are updated from ContentService::publishVersion()
*after* ContentHandler::publish() has been called, the PostPublish
FieldStorageEvent can not be called by the ContentHandler.

It also updates the chain call by making sure the ContentHandler method
only fetches the content from the database if an event has updated it.