Remove EZP-21294: UrlAlias Criterion
This criterion has not been implemented but have existed as interface since 5.0, however it has become clear we can
not support such a criterion on storage engines like Solr because of complex UrlAlias language handling.
So lookup by UrlAlias should first be looked up to get Location Id, and that should be used for used with Location or
Subtree criterion.