EZP-22318: Tweaked fragment framework
Side effect when having a cookie path by SiteAcces (hence at runtime), when using a URI matcher, when you do an ESI or Hinclude,
you'll lose the SiteAccess part in the URI as the fragment path is configured per application, and not accessible at runtime.

Solution here is to ensure that the SiteAccess URI part is correctly reflected when generating the fragment path,
and to ensure the fragment listener accepts it.

As for the FragmentListener, a custom factory has been added to examinate the current pathinfo.