Removed PHP 5.3 support
PHP 5.3 is no longer supported by PHP.net, and we expect new version
of Redhat to arrive before we release eZ Publish 5.4 in the fall which
this change wil be part of.

In the meantime we would like to start to take advantage of PHP 5.4 featurs
for eZ Publish 5.4.

We will move further to PHP 5.5 when Debian comes out with new release in 2015.