Update AliasGenerator.php
Hello,

I have a big problem with image variation. When Original image is not found but is register in content fieldType image , I have error  "Trying to run recursive callback in legacy kernel! Inception!" and I have inifinite message loop in the file log.
Maybe when if it call if a runCallBack is made it is essential to make a return
May be