[Integration Test] Fix wrong assumtion in SubtreeLimitationTest->testSubtreeLimitationForbid
This test was passing wrongly as a byproduct of parentId property on UserGroup being loaded
via LocationService. In this case user should have had access to content as subtree of user
group 13 has exactly same path as the subtree used for limitation, while parent location fell
outside of the limitation causing the exception the test where expecting.

Now parentId property is loaded via SPI and don't have this issue, so test has been
changed to limit the rights to members subtree instead, where editor is not placed.