Fix EZP-23971: If a user is removed while logged in, session for that user will throw an exception
When reloading security token from the session, repository user is not
refreshed correctly (i.e. not re-fetched). Consequence is that if the
user is modified in the backend (e.g. e-mail changed or user deleted),
it's not reflected for the currently logged in user.

This patch fixes it. If user isn't found, a `UsernameNotFoundException`
will be thrown, allowing Symfony `ContextListener` (firewall listener
loading security info from the session) to reset the security token and
kick the user out.