Reflect eZ Platform for behat tests and Readme
Further changes to the readme is aimed at being done in followup PR's.
For instance missing things is moving the description of Legacy Stack
to Legacy Bridge Readme, and API and SPI parts as expanded readmes in their
respective directories. In addtion spec future directory structure and
components split is missing.