[WIP][VersionState] Possible interface for VersionStates
Skipped groupes as it does not seem to make sense for a version to be in two
"workflows" at the same time.
However being in different workflows might make sense (one for articles and different one for blogs).

But open for feedback, not having groups might prevent some advance use cases so might make sense to keep.