REST: Load users by login/email + verify existence
Story: EZP-24820

GET /user/users is changed to return a 404 when there are no results.

If we don't do that, HEAD and GET won't behave the same way, which conflicts
with the HTTP spec.

It is also inconsistent with Symfony's HTTP cache behaviour: upon a miss, the
request is cloned, and forwarded as a GET one. The reasoning is that since GET
and HEAD will have the same behaviour, sending a GET will make sure we can cache
the actual request's body.