Updated the PreviewController for location_view changes
If an existing location_view rule uses a custom controller, we can't
transform the rule to a content_view one (the viewLocation and viewContent
method signatures differ).

This uses the Location ViewProviders to verify if there is such a rule.
If there is, it uses a viewLocation request instead of a viewContent one.

Catch subrequest exceptions in PreviewController

If an exception occurs during the rendering of the preview of a
location that uses a custom controller, display a particular message
explaining that this error case is probably cause by the usage of
the deprecated location view.

Also adds a CompilerPass for ezpublish.view_provider

Added the tag to the existing view_providers.
The existing, per type, tags are still parsed by their existing
compiler passes.