Made PreviewController use viewLocation if custom controller
If an existing location_view rule uses a custom controller, we can't
transform the rule to a content_view one (the viewLocation and viewContent
method signatures differ).

This uses the Location ViewProviders to verify if there is such a rule.
If there is, it uses a viewLocation request instead of a viewContent one.