Refactored the ViewManager
It is kept for BC, as it has been documented as usable by custom controller actions.

Based on the passed value object + view type + params, it will create the View object,
configure it, and pass it for rendering.

It uses the newly added viewConfigurator instead of being injected the viewProviders directly.