Refactored the ViewControllerListener
It is now unique for all View controllers.
Each type of view controller (Content, Block) has its own ViewBuilder:
ContentViewBuilder, BlockViewBuilder, registered with the ViewBuilderRegistry.

The ViewControllerListener will match requests based on the controller,
and use, if any, the matching ViewBuilder to build the View object.

It will then set the request's properties so that the View can be processed.