Bug when using URIElement value grater than 1
In the matcher we end up with an array so the (int) cast will always be 1. 
Not useful for when wanting to use URIElement > 1.
We should always take the value of the config.

A voir avec Florent Huck qui est intervenu sur le sujet en tant que support pour eZPartner SQLI Paris.