[API] Introduce typed SearchResults
Allows user of api to have better type hinting on what kind
of values he has access to.

Implies we will need to do simialr in SPI.
And for SPI LocationSearchHit there would also be option to add a contentInfo property that
Solr can prefill to get rid of db lookups on findLocations, like already on findContentInfo.