Fix EZP-25470: travis fails to install solr package
Composer does not resolve self.version for ezpublish-kernel
in pull request branches against 6.1. Only happens in case
of cyclic dependencies: we try to install ezplatform-solr-search-engine,
that requires ezpublish-kernel, from ezpublish-kernel.

Fixed by renaming the local travis branch to 6.1 so that
requirements can be satisfied. Needs to be adapted to other
stable branches.