Fix EZP-25505: match siteaccess in user hash generation
User hash generation must match the siteaccess to operate properly.
This removes the test that excluded this route from being matched.

The Symfony reverse proxy will use the original request that is passed
as a request attribute.

Varnish will NOT be able to match a siteaccess using an URI. But this
would only be an issue if the URI is required to match a repository.
This is easily prevented by enforcing host/port based siteaccess matching.