Fix EZP-26179: REST session refresh/delete must throw a 404 if session has expired
According to the specifications, /api/ezp/v2/user/sessions/<id>/refresh must
throw a 404 if the session doesn't exist / has expired.

Because of the CSRF token check in the CsrfListener, a 401 is thrown because
the csrf token verification fails, as the session has expired, and we don't
have a stored token anymore.

This changes the CsrfListener to skip this route, and explicitly checks in
the refresh controller action if the token has expired.