Attempt to model "models" for the needs in limitations to be able to model the right values
As in if there is a Content module/model we can clearer specify what limitations belong to.
Meaning, limitations like Subtree/CotnentType/Section becomes limitaitons for content model/module, and only accpt relevant values for that.

However outside of Limitations, and in the terms of DDD / MVC it is strange to denote Content as a model, while Location is not (even if it has own service).