Move kernel to version 7.0 and bump requriments to Symfony 3.x & PHP 7.0
To begin with this will start as a `7.0` branch, acting as a sort of `develop` branch.
So: keep on targeting changes towards master or stable branches as normal, only breaking
changes goes just into `7.0` branch

For features that are also relevant to 6.x typically in a bc safe way into master, then
adaptions to clean up 7.x into `7.0` branhc as seperate pr afterwards _(can be tmp commit on original PR during review)_.

In regards to PHP 7 bump:
With RHEL now having access to PHP 7.x packages via RHSCL 2.3,
and Debian 9 "Stretch" also having this, meaning all support
distroes by the time this is stable has PHP 7.0 support.

For new 7.0 only code we can thus start to take advantage of PHP 7.0 features,
and consider it for existing code and API's where applicable.

Note: Both Symfony 4 and newer versions of Doctrine requires PHP 7.1, so once we get closer
to stable version we might strongly consider that as well to take be able to support both.