Move kernel to version 7.0 and bump requriments to Symfony 3.3 & PHP 7.1
To begin with this will start as a `7.0` branch, acting as a sort of `develop` branch.
So: keep on targeting changes towards master or stable branches as normal, only breaking
changes goes just into `7.0` branch

For features that are also relevant to 6.x typically in a bc safe way into master, then
adaptions to clean up 7.x into `7.0` branhc as seperate pr afterwards _(can be tmp commit on original PR during review)_.

In regards to PHP 7.1 bump:
Both Symfony 4 and newer versions of Doctrine requires PHP 7.1. And it is now avaiable on RHEL via RHSCL 3.0,
upcomming Debian 17.10, only major distro missing it right now is Debian where it is currently in testing.
So for Debian users this means either using testing packages or use thirdparty packages of PHP 7.1 or 7.2
until official packages reaches stable channel.

[Composer] Bump PHP requirement to 7.1 to align with other v2 packages (and Sf4 + Doctrine DBAL 2.6)