EZP-24852: Add UserReference support in Authentication/User providers
As we always refresh APIUser on load anyway, changes our symfony User
to not serialize APIUser to session to reduce size of session being
serialized back and forth.

This approch assumed the refresh is needed for security reasons *(as in
check if user is still exists and is enabled, permissions lookup don't
need this)*. However if that is not the case this can be done closer
to how it is described in JIRA issue and aim to deprecate this logic instead.