Fix EZP-27331: Cannot override view parameters
> https://jira.ez.no/browse/EZP-27331

When using the view system, it is now possible to add arbitrary parameters. 
However it is currently possible to *replace* an existing one as 
`BaseView::addParameters()` uses `+= ` assignment and does not expose a setter.
Thus, if a key in `$parameters` already exists in `$this->parameters`, it is ignored.

This patch replaces `+=` by `array_replace()`, like in Symfony `ParameterBag`.