[PHPUnit] Add forward compatibility with PHPUnit 5.x for 6.7LTS/5.4LTS branch usage
Adds a trait to add to Tests _(or base test cases)_ to be able to use `createMock`.

This marks `getMock` as deprecated, this is purely to hint for new test code to avoid it.
It is NOT meant for refactoring all existing 5.4/6.7 test code to use `createMock`, that is
already done in master (>= 1.13) so any attempt to do that will lead to massive conflicts!