EZP-27924: As a Developer I want lazy loading certain API properties
Introduces concept of lazy loaded properties in order to be
able to expose more information on API, especially for use in
templates but also useful other places, withouth causing more
additional load if not used.

This is here used on ContentType->contentTypeGroups as an example,
and since it is often not used when people load content types.

This is done using proxy object of Content Type Group with takes a
Generator and object id as argument in order to load the objects
when needed. Taking advantage of possibility to send argument to
Generators in order to allow passing the generator into a bunch
of objects for bulk loading, meaning once one is accessed all
are loaded using one single SPI call.

Compared to injecting services into value object this:
- Makes sure to allow bulk loading across objects
- Keeps domain logic in mapper instead of spreading it around in value objects
- Continues to keep value objects "clean of logic" as they have always been

Compared to earlier [POC PR](https://github.com/ezsystems/ezpublish-kernel/pull/2094), this:
- Makes sure to return a normal php array avoiding BC break
- Avoids value objects and API consumer having to have knowledge about Generator

*For further reading see doc/specifications/api/lazy_properties.md*