Assume two siteaccesses, `eng` and `chi`, have been configured to support `eng-US` and `chi-CN` locales respectively. If an article at `/chi/news/article_1` only has `chi-CN` translation available, calling `path(ez_route(null, {language: 'eng-US'}))` in the template would output something like `/eng/chi/view/content/...`, which is incorrect because the `/chi` part should not be there.
This issue occurs when `UrlAliasGenerator` has failed to generate the
URL alias for the specific article in the target language, as the
translation in the target language does not exist, so as the fallback
generator, the `DefaultRouter` will be used to generate the internal URL
for the article. However, by default, the `DefaultRouter` always
prepends the siteaccess, which is `chi` in this case, to the genreated
URL, and finally the `UrlAliasGenerator` will also prepend the target
siteaccess 'eng' to the URL.

To fix this issue, a new parameter `ignoreSiteAccess` has been
introducted in the `$parameters` array passed to `DefaultRouter`, which
indicates if the `DefaultRouter` should ignore the stieaccess from the
start of the URL it generates.

| Q | A |
| --- | --- |
| Bug fix? | yes |
| New feature? | no |
| BC breaks? | no |
| Deprecations? | no |
| Tests pass? | yes |
| Fixed tickets | n/a |
| License | GPL v2.0 |
| Doc PR | n/a |