[RFC][Cache] Limit number of locations we tag to not cripple performance
Reported on slack on some install with several hundred locations for a given content, as Symfony Cache is
not designed for such amount of tags it made the whole system immensely slow. eZ is not really made for
this either atm, as we don't really have paging even on listing locations. Which is why this patch
just ignores anything beyond 25 locations.

Todo:
- [ ] check if main location is among the first, as this one is somewhat more important then the others