Part of EZP-29510 allow last version to be removed if draft
This change allows users to remove the only existing version if it is a draft with only `content versionremove` permissions.  This is needed so that in admin ui if a user creates content but then decides to cancel the process they can do so. I believe `content remove` should only apply to published content.