[Travis] Fix failure on Solr with all branches
Attempt at fixing the failure, the test involved is missing a `refreshSearch()` so
attempting to add it to see if that is the culprit.

As for why this pops up, probab ly beacause we now set config slightly different as of 1.5.3/1.5.4,
we now also set auto soft commit time even for CI runs so might be something in regards to that.