[WIP] EZP-28888: Custom cache_service_name can not be set from siteaccess group
This is more of a woraround then a fix.
As we probably have this issues on all siteaccess aware settings with default values.

My guess is that this is regression from #2009
And that we need to look into if default value is injected into each scope here, or logic in config resolver is missing some adaptions to handle fallbacks better now that scopes can be empty.