Available languages should not depend on languages order
Currently translation helper getAvailableLanguages function uses only
the first entry in the languages configuration of a siteaccess

Therefore if we have three site accesses, we need to define the languages
in each of them and with the current language as the first in the array,
although there is no such restriction for the configuration

For such a siteaccess example

    siteaccess:
        list: [en, de, fr, it, admin]
        groups:
            site_group: [en, de, fr, it]
            switzerland: [de, fr, it]

This would work

    en:
        languages: [eng-GB]

    de:
        languages: [ger-DE, fre-FR, ita-IT]

    fr:
        languages: [fre-FR, ger-DE, ita-IT]

    it:
        languages: [ita-IT, fre-FR, ger-DE]

This

    en:
        languages: [eng-GB]

    de:
        languages: [ger-DE, fre-FR, ita-IT]

    fr:
        languages: [ger-DE, fre-FR, ita-IT]

    it:
        languages: [ger-DE, fre-FR, ita-IT]

or more importantly if we define the languages in the sitegroup

    switzerland:
        languages: [ger-DE, fre-FR, ita-IT]
        translation_siteaccesses: [de, fr, it]

the translation helper would only return the first available language in
the list.

Here I have added a test case and the fix for it. Other tests still work
as expected.