IBX-1228: Prevented storing location-path 1 in the persistence cache
For more details see https://issues.ibexa.co/browse/IBX-1228 and https://github.com/ezsystems/ezpublish-kernel/pull/3122

Currently, we store a huge cache set for location path 1 containing an item for each content object. The reason for that is /1/ part in its location_path.

We don't have scenarios for purging location "1" within the system, so we don't need to persist that data. It may result in some memory usage reduction when Redis is in use (estimate saving is at the level of ~5% for the entire cache data).